Add support for dumping tablespaces and create a test filespace
As a tablespace must be created in a particular filespace and a
user tablespace cannot be created in one of the default filespaces,
this commit adds logic to the integration test suite to create
a temporary "test_filespace" filespace that lives for the duration
of the integration tests and is used for all tablespace tests.

Additionally, the query for GetSegmentConfiguration in utils/io.go
was modified to account for having multiple filespaces in the cluster,
and somce miscellaneous formatting/organization changes were made.