Added dependency resolution for tables that inherit other tables
This commit modifies the way Relations are sorted to preserve
dependency ordering.  To do this, it introduces a DependsUpon
field to the Relation struct to store the names of relations
upon which a relation depends, and modifies the Relations.Less
function used in sorting Relations to order by dependency instead
of by name.

Tables that inherit other tables now have their CREATE TABLE
statements dumped with an INHERITS clause; all of the column
definitions are still printed.

Finally, backup.OidFromObjectName now accepts a schema qualifier
for objects with schemas.

Signed-off-by: Karen Huddleston <khuddleston@pivotal.io>