Refactor functions related to querying and constructing table definitions
Previously, the ConstructDefinitionsForTable was called once for each table
to be dumped, and each call to it in turn called 8 other functions that
queried the database.

The overhead of issuing 8 queries per table in the dump can be quite high
for large numbers of tables, so this commit refactors those functions to
retrieve all the necessary information in one query total, not one per table,
and map that information to individual tables on the Go side rather than the
database side.