Add creation and deletion of pipes for use with plugins
Currently, when given the --plugin flag, gpbackup creates
pipes for all metadata and data files, spins off goroutines
to read the output from the pipes and write them back to disk,
and removes the pipes afterwards.  Later stories will expand
this functionality to do something useful with the pipe output,
such as compressing it or sending it to an external storage module.

Signed-off-by: Karen Huddleston <khuddleston@pivotal.io>