Modify catalog queries to function in 4.3
This commit modifies the catalog queries in the queries_*.go files and the
corresponding tests to be compatible with the 4.3 catalog, and skips backing
up and retrieving information for any database objects that do not exist in
or are not supported by GPDB 4.3.

As part of this effort, the commit introduces a GPDBVersion struct and some
associated version checking functions to allow simple validation of the GPDB
version against which gpbackup is run.  It also adds some helper functions to
testutils to enable similar version comparison in the integration tests.

This commit does not modify any of the predata/postdata SQL statement creation
functions, so there is no guarantee that a 4.3 backup taken with this commit
can be restored to either a 4.3 or 5 database.  That effort will be completed
in an upcoming commit.

Signed-off-by: Chris Hajas <chajas@pivotal.io>
Signed-off-by: Karen Huddleston <khuddleston@pivotal.io>