Modify how hyphens-as-NULLs are handled for base type functions
As part of the previous commit making catalog queries work in 4.3,
the queries related to database functions were changed to return "-"
in place of "" for some fields, as pg_type uses "-" in place of NULL
for those fields and there is no regproc-to-text cast in 4.3 that
would allow turning that into an empty string with a CASE statement.

This made working with base types more confusing, so this commit
modifies the queries to return to the old behavior and return "" for
NULL in both 4.3 and 5, and the tests have been modified appropriately.