Include catalog functions in function info map
Because we need to get the arguments for some built-in functions
when backing up 4.3 (e.g. pg_catalog.plpgsql_validator), we need
to include those functions in the info map.  The actual functions
will not be backed up, just used in statements like ALTER FUNCTION
OWNER, COMMENT ON FUNCTION, and the like.