Combine table map and table of contents files
This commit removes the gpbackup_[timestamp]_table_map file and puts the
relevant information into the gpbackup_[timestamp]_toc.yaml file.  This
makes retrieving the information easier and reduces the number of backup
files to track.

This refactor was prompted by a need to store more information about the
data files for each table, specifically to record the order of attributes
in each table.  As table attribute order can change when one table inherits
another, the original order needs to be preserved, and storing it in this
central location makes more sense than storing it in a data file header or
the like.

Signed-off-by: Chris Hajas <chajas@pivotal.io>