Refactor handling of opening and closing files in utils
Previously, System.OpenFile was a direct alias for os.OpenFile, which returns
a *os.File.  Because *os.File and *gbytes.Buffer don't play nicely with each
other, mocking out file opening and closing was inconvenient and no function
that used System.OpenFile for writing could use a *gbytes.Buffer in tests.

This commit splits System.OpenFile into read and write versions which point to
functions that wrap the result of an os.OpenFile call as a utils.ReadCloserAt
(an interface satisfying io.ReadCloser and io.ReaderAt) or an io.WriteCloser,
respectively.  This allows using a *gbytes.Buffer wherever a *os.File was used
and simplifies testing of functions that open and close files.