Genericize adding table of contents metadata entries
Previously, the TOC had separate functions for adding entries for global,
predata, postdata, and statistics metadata, and functions that were used
to print to multiple such files had to be split into multiple functions
(such as having PrintPredataSessionGUCs, PrintPostdataSessionGUCs, etc.)
even if they were identical in functionality otherwise.

This commit adds a metadataEntryMap field to the TOC struct so that it
can automatically determine to which MetadataEntries array a given Entry
should be added based on the filename of the FileWithByteCount to which
the metadata was written.