Restore all metadata through database connection instead of psql
This commit follows up on a previous commit for genericizing addition of
metadata to the table of contents during backup, in order to similarly
genericize the restoration of metadata.

Previously, the predata, postdata, and statistics files were restored via
an external psql call, so they could not be filtered using the TOC.  Now,
these files are read in and executed via the database connection so they
can be filtered the same way.