Improve handling of autogenerated types
Previously, different functions accounted for autogenerated types (array types
created along with base types and base and composite types created along with
relations) in various inefficient and/or incorrect ways to determine whether
a given type should be backed up or counted as a dependency.

Now, there is a single function to retrieve a list of OIDs of all autogenerated
types that can be used to exclude those types from backups and dependency lists
in an efficient, standardized way.