Modify queries to retrieve reference to implicit array types
In the previous commit, we removed the retrieval of implicit array types (those
created as a side effect of the creation of user-defined types) from queries
that retrieve type dependency information.  This fixed several dependency
resolution issues, but caused an issue with functions dependent on tables.

Now, we instead retrieve implicit array type dependency information, but replace
a dependency on an array type with a dependency on the base or composite type of
which it is an array.

Signed-off-by: Chris Hajas <chajas@pivotal.io>