Restrict reformatting of external table options to custom format
A previous commit changed the output of the fmtopts field in pg_exttable
from the format "key 'value' key 'value'" to "key='value', key='value'"
for all external tables.  This format is required for tables in a custom
format, but disallowed for tables in text or CSV format, so this commit
reverts the reformatting for those table types and retains it for tables
in custom format.

Signed-off-by: Chris Hajas <chajas@pivotal.io>