Split GetNonEnumTypes into separate functions for each type
The query in GetNonEnumTypes was one of the longest-running queries in gpbackup
and was performing poorly on large databases.  To solve this, it has been split
into one Get[type]Types function for each kind of user-defined type, as that
allows each individual query to use fewer and smaller JOINs to cut down on their
execution times.

Also, the refactored queries in these new functions are improved in two ways:
- The queries use "EXCEPT [subquery retrieving types to be excluded]" instead of
  "NOT IN [list of OIDs to exclude]", as the EXCEPT clause scales much better.
- The queries use array_agg to gather the attributes for composite types into an
  array so that only one row is returned for each composite type, instead of
  retrieving one row per attribute per composite type and assembling them later;
  there is no degradation in performance, and it removes the need for a separate
  loop through the types array to assemble the composite types.