Add --jobs flag to gprestore to enable parallel restore
When the flag is set, gprestore will open N parallel connections to the database
and restore table data and post-data metadata in parallel.  Global and pre-data
metadata and statistics are still restored serially.

When the flag is not set, gprestore defaults to 2 connections so there is always
some amount of parallelism in the restore.

Signed-off-by: Chris Hajas <chajas@pivotal.io>