Set session-level GUCs before data restore
Previously, gprestore was not setting session-level GUCs when restoring data,
which could possibly cause data to be restored incorrectly, and no metadata
files were created for a data-only backup.  Now, a data-only backup creates a
pre-data metadata file to store the session GUCs, and the session GUCs are set
when restoring data.