Add support for single-level external partitions
This commit enables gpbackup to back up partition tables with one or more
external partitions, whether partitioned by list or by range.  If the
--leaf-partition-data flag is given, all non-external partitions will be
backed up, otherwise the entire partition table will be skipped and a
message will be printed telling the user to use that flag to back up the
partition table's data.

Support for multi-level partition tables with external partitions will be
added in a later commit.

Signed-off-by: Karen Huddleston <khuddleston@pivotal.io>