Change function for operator test to return type BOOLEAN
Master is now more strict about which types of operators are allowed to
have which properties. Several of the operator options are only
supported with a boolean type so we will change the type of the function
to boolean for the test in order to continue testing that all the
properties work.