Introduce segment agent for single-data-file backup
When backing up to a remote filesystem, opening many file
handles can experience significant performance degradation.  To avoid this,
instead of re-opening the single data file for successive appends or reads
we create an intermediate pipe and tail the data into and out of the pipe,
so that we only incur the file-opening overhead once.

As a consequence, we now perform compression between the
intermediate pipe and the data file, rather than handling it in the COPY
TO PROGRAM call.

Author: Karen Huddleston <khuddleston@pivotal.io>
Author: Chris Hajas <chajas@pivotal.io>
Author: Jamie McAtamney <jmcatamney@pivotal.io>