Move timestamp and lock file generation earlier (#17)
Previously, we were generating the timestamp and the associated lock
file after we had validated the filter lists. Since this validation
requires querying the database, it can take long enough that the
timestamp generated could be over a second later than when the
user started the backup. Now, we generate the timestamp and
lock file before this validation closer to the beginning of the backup.

Author: Chris Hajas <chajas@pivotal.io>