Set gp_default_storage_options session GUC during restore (#18)
Previously, when creating tables during the restore, the storage options
were inherited from the database setting at the time of the backup. However,
this would be incorrect if a table was initially created with no storage
options, as the database GUC would cause storage options to be added. Instead,
we set a session level GUC so that the table is created with the storage
options at the time of table creation.

Author: Chris Hajas <chajas@pivotal.io>