Add interrupt handler to gpbackup and gprestore
To ensure that the cluster is not left in a bad state if someone aborts a backup
or restore with Ctrl-C, this commit moves cleanup-related functions that were
previously called with defer to a DoCleanup function and sets up a signal handler
goroutine to call DoCleanup and exit if an interrupt is received.  DoTeardown
and several other functions are modified to account for a possible abort in the
middle of the function.

Author: Jamie McAtamney <jmcatamney@pivotal.io>