Harden gpbackup -single-data-file to handle unexpected pipe removal (#34)
Previously, if a pipe was removed during gpbackup with -single-data-file, the
backup would continue and report no error. Additionally, the data would end up
getting written to a file, and then removed. The restore would then fail as the
backup was corrupted. Now, we check that the pipe file exists before each
COPY and report an error if it does not exist or is not a pipe.

This adds ~1s per 1k tables to the backup duration.

Author: Chris Hajas <chajas@pivotal.io>