Return to go get ginkgo
go get automatically fetches and installs source code and binaries,
unlike dep ensure, which does not do the installation. The vendor
directory created by go dep is not in our GOPATH so it is not compatible
with go install. Ginkgo is also only used for testing, not for the actual
utility, so for now we feel it is safe to simply pull the latest version.

Author: Chris Hajas <chajas@pivotal.io>
Author: Karen Huddleston <khuddleston@pivotal.io>