Use pg_database_size to get database size instead of gp_toolkit
gp_toolkit does not exist by default in all 4.3 databases (notably
template0 which we use to create our databases when the -createdb flag
is set), therefore we can't rely on it always being there for us to
query. pg_database_size is from postgres and will be present in all
supported versions of GPDB, so we will use that instead.

Authored-by: Karen Huddleston <khuddleston@pivotal.io>