Allow users to include/exclude sequence and views in backup
These can be specified with the --{in,ex}clude-table{,-file} flags.

One behavior change was that it now behaves like pg_dump in the case
where a sequence is owned by a table. Previously, sequences were
excluded/included along with the table, but now they are treated
separately and must be specified explicitly.
The ALTER SEQUENCE OWNER statements are also not backed up if the owning
table is not included.

For the end to end tests:
 * added more coverage for new filtering features
 * reorganized into context blocks for easier reading and focusing

Co-authored-by: Amil Khanzada <akhanzada@pivotal.io>
Co-authored-by: Karen Huddleston <khuddleston@pivotal.io>