Querying AO Incremental metadata, updating TOC
Accumulates modcount and lastDDLTimestamp for AO tables
To get modcount, we first find the aoseg table FQN corresponding to the
ao table. Then from that aoseg table we query the modcount.
To get the lastDDLTimestamp, or the last timestamp at which a
CREATE/ALTER/TRUNCATE has been executed on the ao table, we query
pg_stat_last_operation.
Note: Column-oriented tables are also handled

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>
Co-authored-by: Karen Huddleston <khuddleston@pivotal.io>