Fix backup of external tables with --single-data-file
During the gpbackup --single-data-file refactor to use an agent
approach, the oid list created for the master included external tables,
while the oid list on the segments did not. This caused the backup to
fail as the pipe names did not match (the segment created a pipe with
the name of a regular table, whereas the master executed a COPY TO
command to a pipe with an external table). This only occurred when
--single-data-file was used.

Authored-by: Chris Hajas <chajas@pivotal.io>