Restore from restore plans for both backup types (#97)
* Restore from restore plans for both backup types

Iterate through timestamps found in restore plan of specified backup
and restore table data corresponding to each timestamp.
Also, ensure that if a restore is done on a legacy backup (having a
missing restore plan), a default restore plan is supplied. Here we
have assumed that such a legacy backup can only be a full backup.
Thus, the default restore plan would always be a single entry with
a list of all tables backed up in the full backup.

Co-authored-by: Chris Hajas <chajas@pivotal.io>
Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>
Co-authored-by: Karen Huddleston <khuddleston@pivotal.io>

* Fixed filtering for case : empty restore plan

The restorePlanTableFQNs IncludeSet was not declared to have
AlwaysMatchingFilter to be false. This caused an include check to always
return true if the include set was empty. Added unit test to capture
this scenario

Authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>

* End to end tests for incremental backup

Authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>