Properly quote roles for GRANT statements
Mixed case roles were not being properly quoted in GRANT statements.
Rolenames that contain special characters or capital letters are only
sometimes quoted in the relacl field of pg_class.

We are now capturing the quoted rolename when it is present there,
or replacing the rolename with the quoted one, pulled from a separate
query, when it was not already quoted.

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>
Co-authored-by: Chris Hajas <chajas@pivotal.io>
Co-authored-by: Karen Huddleston <khuddleston@pivotal.io>