Remove panics from DoTeardown in backup and restore
The recover statement that catches panics from other places in the code
is at the top of DoTeardown. If anything in DoTeardown errors out, it
will occur after the recover and will not be caught so we will get a
messy stacktrace. Therefore, DoTeardown should not have anything that panics.

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>
Co-authored-by: Karen Huddleston <khuddleston@pivotal.io>