Fixed metadata-only with plugin
GetFPInfoList determined fpInfoList using the restore plan. When
metadata-only was specified with a plugin, the restore plan was
empty and thus the fpInfoList returned from GetFPInfoList was
empty. Instead, this should have contained the target backup
fpInfo. This prevented the TOC from getting downloaded on restore.
Also, hardened against this with an end-to-end test.

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>
Co-authored-by: Lisa Oakley <loakley@pivotal.io>