Change Compression to PipeThroughProgram
Since the compression object can use either cat or gzip depending on
whether we want to use compression or not, we want to change the name to
be more ambiguous so people won't think that it is always using gzip.

Also changed backupFile and pluginCommand to destination(ToWrite/ToRead) and
(sendTo/readFrom)DestinationCommand.

Co-authored-by: Karen Huddleston <khuddleston@pivotal.io>
Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Chris Hajas <chajas@pivotal.io>