Create next pipe before creating destination backup file for --single-data-file
Previously, backup with --single-data-file and --backup-dir on a slow
filesystem such as BoostFS would fail if the first table in the backup
was small enough to fit in the pipe buffer accross all the segments.
COPY ON SEGMENT would progress past the first table (since it fit in the
pipes' buffers) and move to the subsequent table. However, if creating the
destination backup file took relatively long, COPY would fail since the
subsequent pipe had not yet been created.

We now create the subsequent pipe before creating the backup file.
Additionally, since a reader is only attached to the first pipe, this
issue will not arise on subsequent tables.

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Chris Hajas <chajas@pivotal.io>