Fix: Add ability to restore individual leaf partition tables
Previously, when attempting to restore a leaf partition with
--include-table, restore would fail since the root partition table DDL
was not being restored. Now, we will restore the DDL of the root
partition table when restoring any part of the partition table.

Additionally, we modify the restore validation to check both the
metadata entries and data entries to allow restoring leaf partitions.

Co-authored-by: Chris Hajas <chajas@pivotal.io>
Co-authored-by: Kevin Yeap <kyeap@pivotal.io>