Improve logs around file upload
There is a portion of time between when the GUI progress bar completes
vs when all backup data is completely uploaded via plugin. This "pause"
makes it seems like gpbackup is hanging. We added and improved the
logging around this area to better reflect what gpbackup is doing during
this time.

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Chris Hajas <chajas@pivotal.io>