Set backup and restore lock_timeout GUC
Set lock_timeout to infinite (0) in GPDB 6+ so backup and restore will not timeout
when attempting to aquire locks.

This scenario could occur if a database has a default lock timeout of 1 second, and
gpbackup tries to aquire an access share lock on a table while a DDL
operation is occuring on that table.

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Chris Hajas <chajas@pivotal.io>