Do not backup indexes that are not valid or not ready (#152)
Dumping invalid indexes can cause problems during restore. For example,
if the reason the index creation failed was because it tried to enforce
a uniqueness condition not satisfied by the table's data.  Also, if the
index creation is still in progress, we consider it to be an uncommitted 
DDL change, which we wouldn't want to backup anyway.

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Chris Hajas <chajas@pivotal.io>