Fix restoring databases that contain special characters
Previously, special characters in database names were not handled
correctly during restore because we weren't unquoting and unescaping double quotes from
database names read from the config file, which is in an escaped
quote_ident form. We now unquote this database for use in the
connection.

Additionally, we escape single quotes when necessary during queries.

Co-authored-by: Chris Hajas <chajas@pivotal.io>
Co-authored-by: Kevin Yeap <kyeap@pivotal.io>