Fix restore validation to consider all filtering
We were only considering the --include-table option before and would
assume all tables in the backup set were being restored otherwise. This
could result in validation failing because a table exists in the
database even when it wouldn't really be restored because of other
filtering options.

We now construct a correct list of tables that we intend to restore by
considering all filtering options before checking if they are in the
database or not.

Co-authored-by: Karen Huddleston <khuddleston@pivotal.io>
Co-authored-by: Nadeem Ghani <nghani@pivotal.io>