Provide more information about agent errors after data restore
Previously, we would only check for, and log, an error with a helper
agent if the main process COPY command had not encountered an error
first. Now that our helper agents are creating an error indicator file
everytime they fail, we can check for this file from the main process
and reliably let the user know if the agent had a problem.

Authored-by: Karen Huddleston <khuddleston@pivotal.io>