Fix backup filter validation inconsistencies
When backing up with a nonexistent table and schema using the
--exclude-table/--exclude-schema flag gpbackup would fatal. We now allow
this by removing the validation of the --exclude-table and
--exclude-schema flag as users may want to exclude tables/schemas that
may or may not exist at the time of backup.

When backing up with filtered schemas or tables that do not exist in the
database using --include-schema or --include-table, gpbackup will fail
to validate. This behavior is not intended. Now we properly validate
filter options

Co-Authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-Authored-by: Chris Hajas <chajas@pivotal.io>