Refactor set implementation (#178)
Previously, we were using sets for two purposes:
  1) As an include or exclude set to match table/schemas that were to be
  included or excluded from the backup or restore
  2) As a regular set

The first case uses a modified implementation, where an empty set would
always return that a object exists. This was done so that the default
behavior would backup/restore all objects. However, in cases where a
generic set without the property was wanted, it could lead to confusing
behavior.

This introduces a NewSet constructor and replaces instances that should
have been using a regular set with this call.

Additionally, this removes the Add and Delete set functionality since
they weren't being used.

Authored-by: Chris Hajas <chajas@pivotal.io>