Fix restore filtering for sequences
We want users to be able to specify sequences to restore with --include-table
and allow them to restore even if the owning table of the sequence is not
being restored (maybe the table is already in the database). We also
only want to restore the ALTER SEQUENCE .. OWNER TO .. statement if both
the table and sequence are being restored.

Authored-by: Karen Huddleston <khuddleston@pivotal.io>