Set session GUC standard_conforming_strings to 1
This is important because while this is on by default in GPDB5+, it is
off by default in GPDB4. Having this off causes quote_ident to behave
differently because backslashes are treated as escape characters rather
than literals. We need a backslash in a sample like public."foo\bar" to
be treaded as a literal.

Co-authored-by: Trevor Yacovone <tyacovone@pivotal.io>
Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Larry Hamel <lhamel@pivotal.io>