Add 6X jobs to gpbackup pipeline
With the addition of the 6X_STABLE branch, we must ensure that there is
consistency between the naming of jobs and the branches they track. The
'master' jobs now point to a binary and src that correspond to the
master branch of GPDB, while the newly-added 6X jobs use resources
pulled from the 6X_STABLE branch.

Besides job additions, we refactored jobs(GPDB6,master) in two major ways:
 - They now use local clusters, rather than CCP
 - They now test against RHEL6 & RHEL7 in parallel, rather than just 6

* Add GPDB6 job
* Add scale-6x-stable; Add 'master' resources
* Rename jobs to remove the 'integrations-' prefix & '-stable' suffix
* Remove job 'units'
    * This job is not necessary, as unit tests are being ran in surrounding
jobs.

Co-authored-by: Trevor Yacovone <tyacovone@pivotal.io>
Co-authored-by: Larry Hamel <lhamel@pivotal.io>
Signed-off-by: Larry Hamel <lhamel@pivotal.io>
Signed-off-by: Trevor Yacovone <tyacovone@pivotal.io>
Signed-off-by: Larry Hamel <lhamel@pivotal.io>
Signed-off-by: Trevor Yacovone <tyacovone@pivotal.io>