Add new Makefile target end_to_end_without_install
Given that we are now precompiling binaries in our pipeline, we do not
want to recompile prior to executing the end_to_end test suite.

We add a new Makefile target end_to_end_without_install which will be
used for CI only where we don't want to overwrite the binary that has
been installed by gppkg.

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Trevor Yacovone <tyacovone@pivotal.io>