Set extra_float_digits session GUC during backup (#249)
We need to set extra_float_digits GUC when backing up real, double
precision, and geometric data types to preserve the exact data. Without
the GUC, only up to 6 precision points is backed up which is not enough
for Postgres/Greenplum to reconstruct the proper value.

The GUC max value was increased from 2 to 3 in Postgres 9.0. This means
that Greenplum 6 will have the value set to 3 while versions before that
will have max value 2.

End-to-end tests to check the backup functionality for reals are also added.

Co-authored-by: Junkeun Yi <juyi@pivotal.io>
Co-authored-by: Jimmy Yih <jyih@pivotal.io>
Co-authored-by: Shivram Mani <smani@pivotal.io>