Fix concourse bug: cannot specify sles image in pipeline yml
For some unknown reason, Concourse produces a build error when
attempting to generate a container from a sles image when this image is
specified *in* the pipeline yml, rather than in an external task yml file.

Other teams and pipelines have seen this same issue, and the cause
remains unknown.

Authored-by: Trevor Yacovone <tyacovone@pivotal.io>