Fix SIGINT end_to_end test flake
In our production pipelines, we have noticed that gpbackup and/or
gprestore can finish faster than the thread that sleeps before sending
a SIGINT to the gpbackup/gprestore process. Reduce the upper bounds of
the sleep by 0.3 seconds to greatly reduce our chances of hitting this
flake.

We could make these tests deterministic by adding a fault injector
pause during COPY TO/FROM but the randomness to flush out issues is
fine for now.

Co-authored-by: Trevor Yacovone <tyacovone@pivotal.io>