Enable custom-backup-dir usage in end_to_end suite
Previously, we had added a randomly generated string to the backup directory path
without actually creating this directory. The end_to_end suite did not
contain logic to ensure that the provided directory existed – instead,
it would no-op and we would not properly use the flag.

This commit adds a check to ensure the provided directory exists.
Additionally, we have removed the $RANDOM declaration that existed in
the boostfs test task yml, as unique test dirs are now being generated
within the end_to_end suite itself.

Co-authored-by: Trevor Yacovone <tyacovone@pivotal.io>
Co-authored-by: Shivram Mani <smani@pivotal.io>