Report error if gpbackup_helper script fails to write
Previously, gprestore was writing the script to run gpbackup_helper to the
segments using HEREDOC. It was possible for the writing of the script to
fail due to an issue like a full temp directory. Because the next thing
called after the HEREDOC write was a backgrounded gpbackup_helper call,
a HEREDOC write fail would not be detected and gpbackup could hang
indefinitely. This fix correctly reports an error back to gprestore.

Authored-by: Kevin Yeap <kyeap@pivotal.io>