Adjust "a095dd4 - Report error if gpbackup_helper..."
This commit introduced a flake. The primary intent of commit a095dd4 was
to ensure that the HEREDOC script was written properly - however, we
also added logic to check that the backgrounded gpbackup_helper process
had been started successfully.

Our check would occassionally execute _before_ gpbackup_helper was given
enough time to launch, leading to a false error. Rather than add a
'sleep', we have eliminated this check and will rely on gpbackup to
ensure the helper was started properly.

Co-authored-by: Trevor Yacovone <tyacovone@pivotal.io>
Co-authored-by: Lav Jain <ljain@pivotal.io>