Set GUC gp_use_legacy_hashops when restoring GPDB 4/5 backups to GPDB 6+
In GPDB 6, the default hash operators have changed. Because of this,
we must explicitly use the legacy hash operators when restoring GPDB
4/5 tables to a GPDB 6+ cluster.

Added a warning message as well to highlight the GUC usage. It's
recommended that the user would then ALTER TABLE SET DISTRIBUTED BY
their table when possible to use the new default hash operators.

GPDB Reference:
https://github.com/greenplum-db/gpdb/commit/242783ae9f573bb0

Co-authored-by: Lav Jain <ljain@pivotal.io>